Revert the use of a second IO completion port for notifying the main thread of finished tasks
We ran into issues with the combination of MsgWaitForMultipleObjectsEx and PostQueuedCompletionStatus. The main thread would not be notified of finished tasks, depending on when and if, for example, BeginPaint or ShowWindow was called. The only reason we used IOCP for this was to start tasks before we started creating the window, as we would otherwise need a HWND. This has now been replaced with PostThreadMessage, which does not require a HWND. The only significant change with the previous HWND situation is that we handle the task completion message directly inside the message loop, instead of inside the WindowProc.